Restate the Digest test bound without a narrow multiplication
src/tests/testAuthDigest.cc mirrors the universal value-length bound the
Digest credential parser applies before each field's own capacity rule,
and restates it twice so that a divergence from String::RawSizeMaxXXX()
is caught at compile time. The second restatement multiplied 64 by 1024
in int and let the product widen implicitly for the size_t comparison.
Static analysis flags that shape wherever it appears, even where -- as
here -- both operands are literals inside a static_assert() and the
product cannot overflow.

Multiply in size_t so the comparison needs no implicit widening, and
record why in a comment so the cast is not simplified away later. The
asserted value, and every capacity the test pins down, are unchanged.

+++ src/tests/testAuthDigest.cc
@@ -96,7 +96,9 @@ static_assert(sizeof(DigestFieldNames) / sizeof(DigestFieldNames[0]) == Recogniz
 constexpr size_t UniversalValueLengthBound = 65536;
 
 static_assert(UniversalValueLengthBound == ((3 * 64 * 1024 - 1) + 1) / 3);
-static_assert(UniversalValueLengthBound == 64 * 1024);
+// the second restatement multiplies in size_t rather than int so that the
+// comparison needs no implicit widening of a narrower product
+static_assert(UniversalValueLengthBound == static_cast<size_t>(64) * 1024);
 
 /// How DigestFieldCapacities[] constrains a field's value length. Mirrors the
 /// production DigestLengthRule in order, and in membership for every rule a
